Made all Vec3 functions templatized, because I realised I could.
The only thing stopping me before was the need to call type-specific methods like powf, sqrtf, and so on. But I realised that <cmath> declares template versions of these that invoke the right version for the type you pass in. So if we use those, there's nothing to stop us making templates out of all the remaining functions (which is nice, because it means I don't have to write new versions of them if I want to use doubles instead of floats; or if I one day want to use long doubles...).

Next step will be to do the same to Vec2, Vec4 and Ray3.

+++ src/vgl_camera.cpp
@@ -78,9 +78,9 @@ Camera::~Camera()
 void Camera::panBy(float dx, float dy, float dz)
 {
   Vec3f viewVec = _target - _pos;
-  viewVec = rotateX(viewVec, dx / (2 * M_PI));
-  viewVec = rotateZ(viewVec, dz / (2 * M_PI));
-  viewVec = rotateY(viewVec, dy / (2 * M_PI));
+  viewVec = rotateX(viewVec, dx / (float)(2 * M_PI));
+  viewVec = rotateZ(viewVec, dz / (float)(2 * M_PI));
+  viewVec = rotateY(viewVec, dy / (float)(2 * M_PI));
   _target = viewVec + _pos;
 }
 
@@ -88,9 +88,9 @@ void Camera::panBy(float dx, float dy, float dz)
 void Camera::rollBy(float dx, float dy, float dz)
 {
   Vec3f viewVec = _pos - _target;
-  viewVec = rotateX(viewVec, dx / (2 * M_PI));
-  viewVec = rotateZ(viewVec, dz / (2 * M_PI));
-  viewVec = rotateY(viewVec, dy / (2 * M_PI));
+  viewVec = rotateX(viewVec, dx / (float)(2 * M_PI));
+  viewVec = rotateZ(viewVec, dz / (float)(2 * M_PI));
+  viewVec = rotateY(viewVec, dy / (float)(2 * M_PI));
   _pos = viewVec + _target;
 }
